Revert "[LSP] Add support for reporting project level diagnostics in workspace pull (#61819)"
This reverts commit 3cadbf5a023f6262808cd37cb6bf22905bc41edd.
