+++ src/Features/LanguageServer/ProtocolUnitTests/Diagnostics/PullDiagnosticTests.cs
@@ -3,7 +3,6 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Composition;
 using System.Linq;
@@ -29,7 +28,9 @@ using LSP = Microsoft.VisualStudio.LanguageServer.Protocol;
 
 namespace Microsoft.CodeAnalysis.LanguageServer.UnitTests.Diagnostics
 {
-    public class PullDiagnosticTests : AbstractPullDiagnosticTestsBase
+    using DocumentDiagnosticPartialReport = SumType<SumType<FullDocumentDiagnosticReport, UnchangedDocumentDiagnosticReport>, DocumentDiagnosticPartialResult>;
+
+    public class PullDiagnosticTests : AbstractLanguageServerProtocolTests
     {
         #region Document Diagnostics
 
@@ -285,6 +286,22 @@ class A {";
             Assert.Equal("CS1513", results.Single().Diagnostics.Single().Code);
             Assert.Equal(1, results.Single().Diagnostics.Single().Range.Start.Line);
         }
+
+        private static async Task InsertTextAsync(
+            TestLspServer testLspServer,
+            Document document,
+            int position,
+            string text)
+        {
+            var sourceText = await document.GetTextAsync();
+            var lineInfo = sourceText.Lines.GetLinePositionSpan(new TextSpan(position, 0));
+
+            await testLspServer.InsertTextAsync(document.GetURI(), (lineInfo.Start.Line, lineInfo.Start.Character, text));
+        }
+
+        private static Task OpenDocumentAsync(TestLspServer testLspServer, Document document) => testLspServer.OpenDocumentAsync(document.GetURI());
+
+        private static Task CloseDocumentAsync(TestLspServer testLspServer, Document document) => testLspServer.CloseDocumentAsync(document.GetURI());
 
         [Theory, CombinatorialData]
         public async Task TestStreamingDocumentDiagnostics(bool useVSDiagnostics)
@@ -554,10 +571,9 @@ class B {";
 
             var results = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics);
 
-            Assert.Equal(3, results.Length);
+            Assert.Equal(2, results.Length);
             Assert.Equal("CS1513", results[0].Diagnostics.Single().Code);
             Assert.Empty(results[1].Diagnostics);
-            Assert.Empty(results[2].Diagnostics);
         }
 
         [Theory, CombinatorialData]
@@ -593,8 +609,8 @@ class B {";
 
             var results = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics);
 
-            Assert.Equal(DiagnosticProducingGenerator.Descriptor.Id, results[0].Diagnostics.Single().Code);
-            Assert.Empty(results[1].Diagnostics);
+            var diagnostic = Assert.Single(results.Single().Diagnostics);
+            Assert.Equal(DiagnosticProducingGenerator.Descriptor.Id, diagnostic.Code);
         }
 
         [Theory, CombinatorialData]
@@ -637,10 +653,10 @@ class B {";
 
             var workspaceXml =
 @$"<Workspace>
-    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj1"" FilePath=""C:\CSProj1.csproj"">
+    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj1"">
         <Document FilePath=""C:\C.cs"">{csharpMarkup}</Document>
     </Project>
-    <Project Language=""TypeScript"" CommonReferences=""true"" AssemblyName=""TypeScriptProj"" FilePath=""C:\TypeScriptProj.csproj"">
+    <Project Language=""TypeScript"" CommonReferences=""true"" AssemblyName=""TypeScriptProj"">
         <Document FilePath=""C:\T.ts"">{typeScriptMarkup}</Document>
     </Project>
 </Workspace>";
@@ -649,7 +665,7 @@ class B {";
 
             var results = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics);
 
-            Assert.False(results.Any(r => r.TextDocument!.Uri.LocalPath.Contains(".ts")));
+            Assert.True(results.All(r => r.TextDocument!.Uri.LocalPath == "C:\\C.cs"));
         }
 
         [Theory, CombinatorialData]
@@ -671,11 +687,9 @@ class B {";
             // project.
             results = results.Sort((x, y) => x.Uri.ToString().CompareTo(y.Uri.ToString()));
 
-            Assert.Equal(3, results.Length);
-            // Since we sorted above by URI the first result is the project.
-            Assert.Empty(results[0].Diagnostics);
-            Assert.Equal("CS1513", results[1].Diagnostics.Single().Code);
-            Assert.Empty(results[2].Diagnostics);
+            Assert.Equal(2, results.Length);
+            Assert.Equal("CS1513", results[0].Diagnostics.Single().Code);
+            Assert.Empty(results[1].Diagnostics);
         }
 
         [Theory, CombinatorialData]
@@ -689,25 +703,27 @@ class B {";
 
             var results = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics);
 
-            Assert.Equal(3, results.Length);
+            Assert.Equal(2, results.Length);
             Assert.Equal("CS1513", results[0].Diagnostics.Single().Code);
             Assert.Empty(results[1].Diagnostics);
-            Assert.Empty(results[2].Diagnostics);
 
             testLspServer.TestWorkspace.OnDocumentRemoved(testLspServer.TestWorkspace.Documents.First().Id);
 
             var results2 = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics, previousResults: CreateDiagnosticParamsFromPreviousReports(results));
 
             // First doc should show up as removed.
-            Assert.Equal(3, results2.Length);
+            Assert.Equal(2, results2.Length);
             Assert.Null(results2[0].Diagnostics);
             Assert.Null(results2[0].ResultId);
 
-            // Second and third doc should be changed as the project has changed.
-            Assert.Empty(results2[1].Diagnostics);
+            // Second doc should be changed as the project has changed.
+            Assert.Empty(results[1].Diagnostics);
             Assert.NotEqual(results[1].ResultId, results2[1].ResultId);
-            Assert.Empty(results2[2].Diagnostics);
-            Assert.NotEqual(results[2].ResultId, results2[2].ResultId);
+        }
+
+        private static ImmutableArray<(string resultId, Uri uri)> CreateDiagnosticParamsFromPreviousReports(ImmutableArray<TestDiagnosticResult> results)
+        {
+            return results.Select(r => (r.ResultId, r.Uri)).ToImmutableArray();
         }
 
         [Theory, CombinatorialData]
@@ -721,21 +737,18 @@ class B {";
 
             var results = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics);
 
-            Assert.Equal(3, results.Length);
+            Assert.Equal(2, results.Length);
             Assert.Equal("CS1513", results[0].Diagnostics.Single().Code);
             Assert.Empty(results[1].Diagnostics);
-            Assert.Empty(results[2].Diagnostics);
 
             var results2 = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics, previousResults: CreateDiagnosticParamsFromPreviousReports(results));
 
-            Assert.Equal(3, results2.Length);
+            Assert.Equal(2, results2.Length);
             Assert.Null(results2[0].Diagnostics);
             Assert.Null(results2[1].Diagnostics);
-            Assert.Null(results2[2].Diagnostics);
 
             Assert.Equal(results[0].ResultId, results2[0].ResultId);
             Assert.Equal(results[1].ResultId, results2[1].ResultId);
-            Assert.Equal(results[2].ResultId, results2[2].ResultId);
         }
 
         [Theory, CombinatorialData]
@@ -749,26 +762,23 @@ class B {";
 
             var results = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics);
 
-            Assert.Equal(3, results.Length);
+            Assert.Equal(2, results.Length);
             Assert.Equal("CS1513", results[0].Diagnostics.Single().Code);
             Assert.Empty(results[1].Diagnostics);
-            Assert.Empty(results[2].Diagnostics);
 
             var buffer = testLspServer.TestWorkspace.Documents.First().GetTextBuffer();
             buffer.Insert(buffer.CurrentSnapshot.Length, "}");
 
             var results2 = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics, previousResults: CreateDiagnosticParamsFromPreviousReports(results));
 
-            Assert.Equal(3, results2.Length);
+            Assert.Equal(2, results2.Length);
             Assert.Empty(results2[0].Diagnostics);
             // Project has changed, so we re-computed diagnostics as changes in the first file
             // may have changed results in the second.
             Assert.Empty(results2[1].Diagnostics);
-            Assert.Empty(results2[2].Diagnostics);
 
             Assert.NotEqual(results[0].ResultId, results2[0].ResultId);
             Assert.NotEqual(results[1].ResultId, results2[1].ResultId);
-            Assert.NotEqual(results[2].ResultId, results2[2].ResultId);
         }
 
         [Theory, CombinatorialData]
@@ -782,12 +792,11 @@ class B {";
 
             var results = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics);
 
-            Assert.Equal(3, results.Length);
+            Assert.Equal(2, results.Length);
             Assert.Equal("CS1513", results[0].Diagnostics.Single().Code);
             Assert.Equal(new Position { Line = 0, Character = 9 }, results[0].Diagnostics.Single().Range.Start);
 
             Assert.Empty(results[1].Diagnostics);
-            Assert.Empty(results[2].Diagnostics);
 
             var buffer = testLspServer.TestWorkspace.Documents.First().GetTextBuffer();
             buffer.Insert(0, " ");
@@ -807,8 +816,6 @@ class B {";
 
             Assert.Empty(results2[1].Diagnostics);
             Assert.NotEqual(results[1].ResultId, results2[1].ResultId);
-            Assert.Empty(results2[2].Diagnostics);
-            Assert.NotEqual(results[2].ResultId, results2[2].ResultId);
         }
 
         [Theory, CombinatorialData]
@@ -822,7 +829,7 @@ class B {";
 
             var results = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics);
 
-            Assert.Equal(3, results.Length);
+            Assert.Equal(2, results.Length);
             Assert.Equal("CS1513", results[0].Diagnostics.Single().Code);
             Assert.Equal(new Position { Line = 0, Character = 9 }, results[0].Diagnostics.Single().Range.Start);
 
@@ -842,12 +849,11 @@ class A {";
                 new[] { markup1, markup2 }, BackgroundAnalysisScope.FullSolution, useVSDiagnostics);
 
             var results = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics);
-            Assert.Equal(3, results.Length);
+            Assert.Equal(2, results.Length);
             Assert.Equal(new Uri("C:/test1.cs"), results[0].TextDocument!.Uri);
             Assert.Equal("CS1513", results[0].Diagnostics.Single().Code);
             Assert.Equal(1, results[0].Diagnostics.Single().Range.Start.Line);
             Assert.Empty(results[1].Diagnostics);
-            Assert.Empty(results[2].Diagnostics);
         }
 
         [Theory, CombinatorialData]
@@ -866,11 +872,11 @@ class A {";
 
             var workspaceXml =
 @$"<Workspace>
-    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj1"" FilePath=""C:\CSProj1.csproj"">
+    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj1"">
         <Document FilePath=""C:\A.cs"">{markup1}</Document>
         <ProjectReference>CSProj2</ProjectReference>
     </Project>
-    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj2"" FilePath=""C:\CSProj2.csproj"">
+    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj2"">
         <Document FilePath=""C:\B.cs"">{markup2}</Document>
     </Project>
 </Workspace>";
@@ -882,9 +888,9 @@ class A {";
             // and a diagnostic in B.cs since it is missing the class name.
             var results = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics);
             AssertEx.NotNull(results);
-            Assert.Equal(4, results.Length);
+            Assert.Equal(2, results.Length);
             Assert.Equal("CS0246", results[0].Diagnostics.Single().Code);
-            Assert.Equal("CS1001", results[2].Diagnostics.Single().Code);
+            Assert.Equal("CS1001", results[1].Diagnostics.Single().Code);
 
             // Insert B into B.cs via the workspace.
             var caretLocation = testLspServer.GetLocations("caret").First().Range;
@@ -896,15 +902,15 @@ class A {";
             var previousResultIds = CreateDiagnosticParamsFromPreviousReports(results);
             results = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics, previousResults: previousResultIds);
             AssertEx.NotNull(results);
-            Assert.Equal(4, results.Length);
+            Assert.Equal(2, results.Length);
 
             // Verify diagnostics for A.cs are updated as the type B now exists.
             Assert.Empty(results[0].Diagnostics);
             Assert.NotEqual(previousResultIds[0].resultId, results[0].ResultId);
 
             // Verify diagnostics for B.cs are updated as the class definition is now correct.
-            Assert.Empty(results[2].Diagnostics);
-            Assert.NotEqual(previousResultIds[2].resultId, results[2].ResultId);
+            Assert.Empty(results[1].Diagnostics);
+            Assert.NotEqual(previousResultIds[1].resultId, results[1].ResultId);
         }
 
         [Theory, CombinatorialData]
@@ -934,15 +940,15 @@ class A {";
 
             var workspaceXml =
 @$"<Workspace>
-    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj1"" FilePath=""C:\CSProj1.csproj"">
+    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj1"">
         <ProjectReference>CSProj2</ProjectReference>
         <Document FilePath=""C:\A.cs"">{markup1}</Document>
     </Project>
-    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj2"" FilePath=""C:\CSProj2.csproj"">
+    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj2"">
         <ProjectReference>CSProj3</ProjectReference>
         <Document FilePath=""C:\B.cs"">{markup2}</Document>
     </Project>
-    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj3"" FilePath=""C:\CSProj3.csproj"">
+    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj3"">
         <Document FilePath=""C:\C.cs"">{markup3}</Document>
     </Project>
 </Workspace>";
@@ -953,13 +959,10 @@ class A {";
             // Verify we have a diagnostic in C.cs initially.
             var results = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics);
             AssertEx.NotNull(results);
-            Assert.Equal(6, results.Length);
+            Assert.Equal(3, results.Length);
             Assert.Empty(results[0].Diagnostics);
             Assert.Empty(results[1].Diagnostics);
-            Assert.Empty(results[2].Diagnostics);
-            Assert.Empty(results[3].Diagnostics);
-            Assert.Equal("CS1001", results[4].Diagnostics.Single().Code);
-            Assert.Empty(results[5].Diagnostics);
+            Assert.Equal("CS1001", results[2].Diagnostics.Single().Code);
 
             // Insert C into C.cs via the workspace.
             var caretLocation = testLspServer.GetLocations("caret").First().Range;
@@ -971,24 +974,18 @@ class A {";
             var previousResultIds = CreateDiagnosticParamsFromPreviousReports(results);
             results = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics, previousResults: previousResultIds).ConfigureAwait(false);
             AssertEx.NotNull(results);
-            Assert.Equal(6, results.Length);
+            Assert.Equal(3, results.Length);
 
             // Verify that new diagnostics are returned for all files (even though the diagnostics for the first two files are the same)
             // since we re-calculate when transitive project dependencies change.
             Assert.Empty(results[0].Diagnostics);
             Assert.NotEqual(previousResultIds[0].resultId, results[0].ResultId);
+
             Assert.Empty(results[1].Diagnostics);
             Assert.NotEqual(previousResultIds[1].resultId, results[1].ResultId);
 
             Assert.Empty(results[2].Diagnostics);
             Assert.NotEqual(previousResultIds[2].resultId, results[2].ResultId);
-            Assert.Empty(results[3].Diagnostics);
-            Assert.NotEqual(previousResultIds[3].resultId, results[3].ResultId);
-
-            Assert.Empty(results[4].Diagnostics);
-            Assert.NotEqual(previousResultIds[4].resultId, results[4].ResultId);
-            Assert.Empty(results[5].Diagnostics);
-            Assert.NotEqual(previousResultIds[5].resultId, results[5].ResultId);
         }
 
         [Theory, CombinatorialData]
@@ -1007,10 +1004,10 @@ class A {";
 
             var workspaceXml =
 @$"<Workspace>
-    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj1"" FilePath=""C:\CSProj1.csproj"">
+    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj1"">
         <Document FilePath=""C:\A.cs"">{markup1}</Document>
     </Project>
-    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj2"" FilePath=""C:\CSProj2.csproj"">
+    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj2"">
         <Document FilePath=""C:\B.cs"">{markup2}</Document>
     </Project>
 </Workspace>";
@@ -1022,11 +1019,9 @@ class A {";
             // and a diagnostic in B.cs since it is missing the class name.
             var results = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics);
             AssertEx.NotNull(results);
-            Assert.Equal(4, results.Length);
+            Assert.Equal(2, results.Length);
             Assert.Equal("CS0246", results[0].Diagnostics.Single().Code);
-            Assert.Empty(results[1].Diagnostics);
-            Assert.Equal("CS1001", results[2].Diagnostics.Single().Code);
-            Assert.Empty(results[3].Diagnostics);
+            Assert.Equal("CS1001", results[1].Diagnostics.Single().Code);
 
             // Insert B into B.cs via the workspace.
             var caretLocation = testLspServer.GetLocations("caret").First().Range;
@@ -1038,19 +1033,15 @@ class A {";
             var previousResultIds = CreateDiagnosticParamsFromPreviousReports(results);
             results = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics, previousResultIds);
             AssertEx.NotNull(results);
-            Assert.Equal(4, results.Length);
+            Assert.Equal(2, results.Length);
 
             // Verify the diagnostic result for A.cs is unchanged as A.cs does not reference CSProj2.
             Assert.Null(results[0].Diagnostics);
             Assert.Equal(previousResultIds[0].resultId, results[0].ResultId);
-            Assert.Null(results[1].Diagnostics);
-            Assert.Equal(previousResultIds[1].resultId, results[1].ResultId);
 
             // Verify that the diagnostics result for B.cs reflects the change we made to it.
-            Assert.Empty(results[2].Diagnostics);
-            Assert.NotEqual(previousResultIds[2].resultId, results[2].ResultId);
-            Assert.Empty(results[3].Diagnostics);
-            Assert.NotEqual(previousResultIds[3].resultId, results[3].ResultId);
+            Assert.Empty(results[1].Diagnostics);
+            Assert.NotEqual(previousResultIds[1].resultId, results[1].ResultId);
         }
 
         [Theory, CombinatorialData]
@@ -1069,11 +1060,11 @@ class A {";
 
             var workspaceXml =
 @$"<Workspace>
-    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj1"" FilePath=""C:\CSProj1.csproj"">
+    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj1"">
         <Document FilePath=""C:\A.cs"">{markup1}</Document>
         <ProjectReference>CSProj2</ProjectReference>
     </Project>
-    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj2"" FilePath=""C:\CSProj2.csproj"">
+    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj2"">
         <Document FilePath=""C:\B.cs"">{markup2}</Document>
     </Project>
 </Workspace>";
@@ -1085,9 +1076,9 @@ class A {";
             // and a diagnostic in B.cs since it is missing the class name.
             var results = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics);
             AssertEx.NotNull(results);
-            Assert.Equal(4, results.Length);
+            Assert.Equal(2, results.Length);
             Assert.Equal("CS0246", results[0].Diagnostics.Single().Code);
-            Assert.Equal("CS1001", results[2].Diagnostics.Single().Code);
+            Assert.Equal("CS1001", results[1].Diagnostics.Single().Code);
 
             // Change and reload the project via the workspace.
             var projectInfo = testLspServer.TestWorkspace.Projects.Where(p => p.AssemblyName == "CSProj2").Single().ToProjectInfo();
@@ -1101,11 +1092,11 @@ class A {";
             results = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics, previousResults: previousResultIds);
 
             AssertEx.NotNull(results);
-            Assert.Equal(4, results.Length);
+            Assert.Equal(2, results.Length);
 
             // The diagnostics should have been recalculated for both projects as a referenced project changed.
             Assert.Equal("CS0246", results[0].Diagnostics.Single().Code);
-            Assert.Equal("CS1001", results[2].Diagnostics.Single().Code);
+            Assert.Equal("CS1001", results[1].Diagnostics.Single().Code);
         }
 
         [Theory, CombinatorialData]
@@ -1124,11 +1115,11 @@ class A {";
 
             var workspaceXml =
 @$"<Workspace>
-    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj1"" FilePath=""C:\CSProj1.csproj"">
+    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj1"">
         <Document FilePath=""C:\A.cs"">{markup1}</Document>
         <ProjectReference>CSProj2</ProjectReference>
     </Project>
-    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj2"" FilePath=""C:\CSProj2.csproj"">
+    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj2"">
         <Document FilePath=""C:\B.cs"">{markup2}</Document>
     </Project>
 </Workspace>";
@@ -1140,9 +1131,9 @@ class A {";
             // and a diagnostic in B.cs since it is missing the class name.
             var results = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics);
             AssertEx.NotNull(results);
-            Assert.Equal(4, results.Length);
+            Assert.Equal(2, results.Length);
             Assert.Equal("CS0246", results[0].Diagnostics.Single().Code);
-            Assert.Equal("CS1001", results[2].Diagnostics.Single().Code);
+            Assert.Equal("CS1001", results[1].Diagnostics.Single().Code);
 
             // Reload the project via the workspace.
             var projectInfo = testLspServer.TestWorkspace.Projects.Where(p => p.AssemblyName == "CSProj2").Single().ToProjectInfo();
@@ -1156,13 +1147,13 @@ class A {";
 
             // Verify that since no actual changes have been made we report unchanged diagnostics.
             AssertEx.NotNull(results);
-            Assert.Equal(4, results.Length);
+            Assert.Equal(2, results.Length);
 
             // Diagnostics should be unchanged as the referenced project was only unloaded / reloaded, but did not actually change.
             Assert.Null(results[0].Diagnostics);
             Assert.Equal(previousResultIds[0].resultId, results[0].ResultId);
-            Assert.Null(results[2].Diagnostics);
-            Assert.Equal(previousResultIds[2].resultId, results[2].ResultId);
+            Assert.Null(results[1].Diagnostics);
+            Assert.Equal(previousResultIds[1].resultId, results[1].ResultId);
         }
 
         [Theory, CombinatorialData]
@@ -1176,15 +1167,15 @@ class A {";
 
             var workspaceXml =
 @$"<Workspace>
-    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj1"" FilePath=""C:\CSProj1.csproj"">
+    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj1"">
         <Document FilePath=""C:\A.cs"">{markup1}</Document>
         <ProjectReference>CSProj2</ProjectReference>
         <ProjectReference>CSProj3</ProjectReference>
     </Project>
-    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj2"" FilePath=""C:\CSProj2.csproj"">
+    <Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj2"">
         <Document FilePath=""C:\B.cs""></Document>
     </Project>
-<Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj3"" FilePath=""C:\CSProj3.csproj"">
+<Project Language=""C#"" CommonReferences=""true"" AssemblyName=""CSProj3"">
         <Document FilePath=""C:\C.cs""></Document>
     </Project>
 </Workspace>";
@@ -1196,7 +1187,7 @@ class A {";
             // and a diagnostic in B.cs since it is missing the class name.
             var results = await RunGetWorkspacePullDiagnosticsAsync(testLspServer, useVSDiagnostics);
             AssertEx.NotNull(results);
-            Assert.Equal(6, results.Length);
+            Assert.Equal(3, results.Length);
             Assert.Equal("CS0246", results[0].Diagnostics.Single().Code);
 
             // Reload the project via the workspace.
@@ -1212,7 +1203,7 @@ class A {";
 
             // Verify that since no actual changes have been made we report unchanged diagnostics.
             AssertEx.NotNull(results);
-            Assert.Equal(6, results.Length);
+            Assert.Equal(3, results.Length);
 
             // Diagnostics should be unchanged as a referenced project was unloaded and reloaded.  Order should not matter.
             Assert.Null(results[0].Diagnostics);
@@ -1221,5 +1212,225 @@ class A {";
         }
 
         #endregion
+
+        /// <summary>
+        /// Helper type to store unified LSP diagnostic results.
+        /// Diagnostics are null when unchanged.
+        /// </summary>
+        private record TestDiagnosticResult(Uri Uri, string ResultId, LSP.Diagnostic[]? Diagnostics)
+        {
+            public TextDocumentIdentifier TextDocument { get; } = new TextDocumentIdentifier { Uri = Uri };
+        }
+
+        private static async Task<ImmutableArray<TestDiagnosticResult>> RunGetDocumentPullDiagnosticsAsync(
+            TestLspServer testLspServer,
+            Uri uri,
+            bool useVSDiagnostics,
+            string? previousResultId = null,
+            bool useProgress = false)
+        {
+            await testLspServer.WaitForDiagnosticsAsync();
+
+            if (useVSDiagnostics)
+            {
+                BufferedProgress<VSInternalDiagnosticReport>? progress = useProgress ? BufferedProgress.Create<VSInternalDiagnosticReport>(null) : null;
+                var diagnostics = await testLspServer.ExecuteRequestAsync<VSInternalDocumentDiagnosticsParams, VSInternalDiagnosticReport[]>(
+                    VSInternalMethods.DocumentPullDiagnosticName,
+                    CreateDocumentDiagnosticParams(uri, previousResultId, progress),
+                    CancellationToken.None).ConfigureAwait(false);
+
+                if (useProgress)
+                {
+                    Assert.Null(diagnostics);
+                    diagnostics = progress!.Value.GetValues();
+                }
+
+                AssertEx.NotNull(diagnostics);
+                return diagnostics.Select(d => new TestDiagnosticResult(uri, d.ResultId!, d.Diagnostics)).ToImmutableArray();
+            }
+            else
+            {
+                BufferedProgress<DocumentDiagnosticPartialReport>? progress = useProgress ? BufferedProgress.Create<DocumentDiagnosticPartialReport>(null) : null;
+                var diagnostics = await testLspServer.ExecuteRequestAsync<DocumentDiagnosticParams, SumType<FullDocumentDiagnosticReport, UnchangedDocumentDiagnosticReport>?>(
+                    ExperimentalMethods.TextDocumentDiagnostic,
+                    CreateProposedDocumentDiagnosticParams(uri, previousResultId, progress),
+                    CancellationToken.None).ConfigureAwait(false);
+                if (useProgress)
+                {
+                    Assert.Null(diagnostics);
+                    diagnostics = progress!.Value.GetValues().Single().First;
+                }
+
+                if (diagnostics == null)
+                {
+                    // The public LSP spec returns null when no diagnostics are available for a document wheres VS returns an empty array.
+                    return ImmutableArray<TestDiagnosticResult>.Empty;
+                }
+                else if (diagnostics.Value.Value is UnchangedDocumentDiagnosticReport)
+                {
+                    // The public LSP spec returns different types when unchanged in contrast to VS which just returns null diagnostic array.
+                    return ImmutableArray.Create(new TestDiagnosticResult(uri, diagnostics.Value.Second.ResultId!, null));
+                }
+                else
+                {
+                    return ImmutableArray.Create(new TestDiagnosticResult(uri, diagnostics.Value.First.ResultId!, diagnostics.Value.First.Items));
+                }
+            }
+
+            static DocumentDiagnosticParams CreateProposedDocumentDiagnosticParams(
+                Uri uri,
+                string? previousResultId = null,
+                IProgress<DocumentDiagnosticPartialReport[]>? progress = null)
+            {
+                return new DocumentDiagnosticParams(new TextDocumentIdentifier { Uri = uri }, null, previousResultId, progress, null);
+            }
+        }
+
+        private static async Task<ImmutableArray<TestDiagnosticResult>> RunGetWorkspacePullDiagnosticsAsync(
+            TestLspServer testLspServer,
+            bool useVSDiagnostics,
+            ImmutableArray<(string resultId, Uri uri)>? previousResults = null,
+            bool useProgress = false)
+        {
+            await testLspServer.WaitForDiagnosticsAsync();
+
+            if (useVSDiagnostics)
+            {
+                BufferedProgress<VSInternalWorkspaceDiagnosticReport>? progress = useProgress ? BufferedProgress.Create<VSInternalWorkspaceDiagnosticReport>(null) : null;
+                var diagnostics = await testLspServer.ExecuteRequestAsync<VSInternalWorkspaceDiagnosticsParams, VSInternalWorkspaceDiagnosticReport[]>(
+                VSInternalMethods.WorkspacePullDiagnosticName,
+                CreateWorkspaceDiagnosticParams(previousResults, progress),
+                CancellationToken.None).ConfigureAwait(false);
+
+                if (useProgress)
+                {
+                    Assert.Null(diagnostics);
+                    diagnostics = progress!.Value.GetValues();
+                }
+
+                AssertEx.NotNull(diagnostics);
+                return diagnostics.Select(d => new TestDiagnosticResult(d.TextDocument!.Uri, d.ResultId!, d.Diagnostics)).ToImmutableArray();
+            }
+            else
+            {
+                BufferedProgress<WorkspaceDiagnosticReport>? progress = useProgress ? BufferedProgress.Create<WorkspaceDiagnosticReport>(null) : null;
+                var returnedResult = await testLspServer.ExecuteRequestAsync<WorkspaceDiagnosticParams, WorkspaceDiagnosticReport?>(
+                    ExperimentalMethods.WorkspaceDiagnostic,
+                    CreateProposedWorkspaceDiagnosticParams(previousResults, progress),
+                    CancellationToken.None).ConfigureAwait(false);
+
+                if (useProgress)
+                {
+                    Assert.Null(returnedResult);
+                    var progressValues = progress!.Value.GetValues();
+                    Assert.NotNull(progressValues);
+                    return progressValues.SelectMany(value => value.Items).Select(diagnostics => ConvertWorkspaceDiagnosticResult(diagnostics)).ToImmutableArray();
+
+                }
+
+                AssertEx.NotNull(returnedResult);
+                return returnedResult.Items.Select(diagnostics => ConvertWorkspaceDiagnosticResult(diagnostics)).ToImmutableArray();
+            }
+
+            static WorkspaceDiagnosticParams CreateProposedWorkspaceDiagnosticParams(
+                ImmutableArray<(string resultId, Uri uri)>? previousResults = null,
+                IProgress<WorkspaceDiagnosticReport[]>? progress = null)
+            {
+                var previousResultsLsp = previousResults?.Select(r => new PreviousResultId(r.uri, r.resultId)).ToArray() ?? Array.Empty<PreviousResultId>();
+                return new WorkspaceDiagnosticParams(identifier: null, previousResultsLsp, workDoneToken: null, partialResultToken: progress);
+            }
+
+            static TestDiagnosticResult ConvertWorkspaceDiagnosticResult(SumType<WorkspaceFullDocumentDiagnosticReport, WorkspaceUnchangedDocumentDiagnosticReport> workspaceReport)
+            {
+                if (workspaceReport.Value is WorkspaceFullDocumentDiagnosticReport fullReport)
+                {
+                    return new TestDiagnosticResult(fullReport.Uri, fullReport.ResultId!, fullReport.Items);
+                }
+                else
+                {
+                    var unchangedReport = (WorkspaceUnchangedDocumentDiagnosticReport)workspaceReport.Value!;
+                    return new TestDiagnosticResult(unchangedReport.Uri, unchangedReport.ResultId!, null);
+                }
+            }
+        }
+
+        private static VSInternalDocumentDiagnosticsParams CreateDocumentDiagnosticParams(
+            Uri uri,
+            string? previousResultId = null,
+            IProgress<VSInternalDiagnosticReport[]>? progress = null)
+        {
+            return new VSInternalDocumentDiagnosticsParams
+            {
+                TextDocument = new LSP.TextDocumentIdentifier { Uri = uri },
+                PreviousResultId = previousResultId,
+                PartialResultToken = progress,
+            };
+        }
+
+        private static VSInternalWorkspaceDiagnosticsParams CreateWorkspaceDiagnosticParams(
+            ImmutableArray<(string resultId, Uri uri)>? previousResults = null,
+            IProgress<VSInternalWorkspaceDiagnosticReport[]>? progress = null)
+        {
+            return new VSInternalWorkspaceDiagnosticsParams
+            {
+                PreviousResults = previousResults?.Select(r => new VSInternalDiagnosticParams { PreviousResultId = r.resultId, TextDocument = new TextDocumentIdentifier { Uri = r.uri } }).ToArray(),
+                PartialResultToken = progress,
+            };
+        }
+
+        private Task<TestLspServer> CreateTestWorkspaceWithDiagnosticsAsync(string markup, BackgroundAnalysisScope scope, bool useVSDiagnostics, bool pullDiagnostics = true)
+            => CreateTestWorkspaceWithDiagnosticsAsync(markup, scope, pullDiagnostics ? DiagnosticMode.Pull : DiagnosticMode.Push, useVSDiagnostics);
+
+        private async Task<TestLspServer> CreateTestWorkspaceWithDiagnosticsAsync(string markup, BackgroundAnalysisScope scope, DiagnosticMode mode, bool useVSDiagnostics, WellKnownLspServerKinds serverKind = WellKnownLspServerKinds.AlwaysActiveVSLspServer)
+        {
+            var testLspServer = await CreateTestLspServerAsync(markup, useVSDiagnostics ? CapabilitiesWithVSExtensions : new LSP.ClientCapabilities(), serverKind);
+            InitializeDiagnostics(scope, testLspServer, mode);
+            return testLspServer;
+        }
+
+        private async Task<TestLspServer> CreateTestWorkspaceFromXmlAsync(string xmlMarkup, BackgroundAnalysisScope scope, bool useVSDiagnostics, bool pullDiagnostics = true)
+        {
+            var testLspServer = await CreateXmlTestLspServerAsync(xmlMarkup, clientCapabilities: useVSDiagnostics ? CapabilitiesWithVSExtensions : new LSP.ClientCapabilities());
+            InitializeDiagnostics(scope, testLspServer, pullDiagnostics ? DiagnosticMode.Pull : DiagnosticMode.Push);
+            return testLspServer;
+        }
+
+        private Task<TestLspServer> CreateTestWorkspaceWithDiagnosticsAsync(string[] markups, BackgroundAnalysisScope scope, bool useVSDiagnostics, bool pullDiagnostics = true)
+            => CreateTestWorkspaceWithDiagnosticsAsync(markups, Array.Empty<string>(), scope, useVSDiagnostics, pullDiagnostics);
+
+        private async Task<TestLspServer> CreateTestWorkspaceWithDiagnosticsAsync(string[] markups, string[] sourceGeneratedMarkups, BackgroundAnalysisScope scope, bool useVSDiagnostics, bool pullDiagnostics = true)
+        {
+            var testLspServer = await CreateTestLspServerAsync(markups, sourceGeneratedMarkups, useVSDiagnostics ? CapabilitiesWithVSExtensions : new LSP.ClientCapabilities());
+            InitializeDiagnostics(scope, testLspServer, pullDiagnostics ? DiagnosticMode.Pull : DiagnosticMode.Push);
+            return testLspServer;
+        }
+
+        private static void InitializeDiagnostics(BackgroundAnalysisScope scope, TestLspServer testLspServer, DiagnosticMode diagnosticMode)
+        {
+            var analyzerReference = new TestAnalyzerReferenceByLanguage(DiagnosticExtensions.GetCompilerDiagnosticAnalyzersMap().Add(
+                InternalLanguageNames.TypeScript, ImmutableArray.Create<DiagnosticAnalyzer>(new MockTypescriptDiagnosticAnalyzer())));
+
+            testLspServer.InitializeDiagnostics(scope, diagnosticMode, analyzerReference);
+        }
+
+        protected override TestComposition Composition => base.Composition.AddParts(typeof(MockTypescriptDiagnosticAnalyzer));
+
+        [DiagnosticAnalyzer(InternalLanguageNames.TypeScript), PartNotDiscoverable]
+        private class MockTypescriptDiagnosticAnalyzer : DocumentDiagnosticAnalyzer
+        {
+            public static readonly DiagnosticDescriptor Descriptor = new DiagnosticDescriptor(
+            "TS01", "TS error", "TS error", "Error", DiagnosticSeverity.Error, isEnabledByDefault: true);
+
+            public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics => ImmutableArray.Create(Descriptor);
+
+            public override Task<ImmutableArray<Diagnostic>> AnalyzeSemanticsAsync(Document document, CancellationToken cancellationToken)
+                => SpecializedTasks.EmptyImmutableArray<Diagnostic>();
+
+            public override Task<ImmutableArray<Diagnostic>> AnalyzeSyntaxAsync(Document document, CancellationToken cancellationToken)
+            {
+                return Task.FromResult(ImmutableArray.Create(
+                    Diagnostic.Create(Descriptor, Location.Create(document.FilePath!, default, default))));
+            }
+        }
     }
 }
